+++ NEWS.md
@@ -3,6 +3,9 @@
 ### Enhancements
 * Added argument `conf_type` to module `tm_g_km` to allow users to set a confidence interval type for median survival time estimation.
 
+### Enhancements
+* Added `denom` argument in `tm_t_binary_outcome` module.
+
 # teal.modules.clinical 0.10.0
 
 ### Enhancements